Refactor: Tests
